+++ src/utils.js
@@ -1,5 +1,3 @@
-const NAMESPACES = ['collection', 'member']
-
-export const reduceNamespaces = (reducer) => NAMESPACES.reduce(reducer, {})
+export const reduceNamespaces = (reducer) => ['collection', 'member'].reduce(reducer, {})
 
 export const forOwn = (obj, iteratee) => Object.keys(obj).forEach((key) => iteratee(obj[key], key))
